+++ NEWS.md
@@ -1,5 +1,10 @@
 # `tidyfr` 0.99
 
+## Changes in 0.99.11
+
+- Use AIDs as `rownames` for the `data.frame` returned by `data` and drop
+  the additional column `"aid"`.
+
 ## Changes in 0.99.10
 
 - Add Rmd for TDFF usage of the metabolomics_p180 data set.